==> code/08_combined_tikz.Rmd <==
---
title: "Results Notebook"
output:
  html_document:
    df_print: paged
---

<!--

*Run* by pressing *Ctrl+Shift+Enter*. 
*Insert Chunk* by pressing *Ctrl+Alt+I*.

-->


```{r}
library(ggplot2)
library(scales)
library(tidyverse)
```

number of iterations per bar: 10000


```{r}
selection = read.csv("../results/06_showcase_results.csv") 

selection$success <- selection$tall_brown - selection$brown_tall

levels(selection$sd_brown) <- c("$\\sigma_\\text{brown}=0.1$","$\\sigma_\\text{brown}=0.2$")
levels(selection$sd_tall) <- c("$\\sigma_\\text{big}=0.25$","$\\sigma_\\text{big}=0.3$")
```



# Mean mean success probability
```{r}
ann_text <- data.frame(lab = "success threshold", success=-0.01, n_obj=factor(8), theta=factor(0.2), sd_brown="$\\sigma_\\text{brown}=0.1$",sd_tall="$\\sigma_\\text{big}=0.25$")
selection$success <- selection$tall_brown - selection$brown_tall

results_plot = selection %>%
  ggplot(aes(x = as.factor(n_obj), y = success, fill = as.factor(theta))) +
  geom_bar(stat = "identity", position = "dodge") + facet_grid(sd_brown ~ sd_tall, scales="free") +
  ylab("difference in mean success probability") +
  xlab("number of objects in context")  +
  scale_fill_brewer(palette="Set1", name = "semantic threshold") +
  expand_limits(y = -0.03) + geom_hline(aes(yintercept= 0), colour= 'black', linetype = "dashed")+ 
  geom_text(data = ann_text, aes(label = lab), fontface = "italic", size = 3, check_overlap = TRUE)+
  theme_bw() +
  theme(legend.position="bottom")

show(results_plot)
```

# Exceptions plot

```{r}

ann_text <- data.frame(lab = "$< 50\\%$", proportion_exceptions=0.47, n_obj=factor(4), theta=factor(0.2), sd_brown="$\\sigma_\\text{brown}=0.1$",sd_tall="$\\sigma_\\text{big}=0.25$")

exceptions_plot = selection %>%
  ggplot(aes(x = as.factor(n_obj), y = proportion_exceptions, fill = as.factor(theta))) +
  geom_bar(stat = "identity", position = "dodge") + facet_grid(sd_brown ~ sd_tall) +
  ylab("proportion of exceptions") +
  xlab("number of objects in context")  +
  scale_fill_brewer(palette="Set1", name = "semantic threshold") +
  expand_limits(y = 0.55) + geom_hline( yintercept = 0) + geom_hline(aes(yintercept= 0.5), colour= 'black', linetype="dashed") +
  geom_text(data = ann_text, aes(label = lab), fontface = "italic", size = 3, check_overlap = TRUE)+
  theme_bw() +
  theme(legend.position="bottom")

show(exceptions_plot +  scale_y_continuous(labels=percent))
```



# Combined graph

```{r}
reverse_scale = 0.25

combined_plot = selection %>%
  ggplot(aes(x = as.factor(n_obj))) +
  geom_col(aes(y = success, fill = as.factor(theta)), position = "dodge") +
  geom_col(aes(y = -reverse_scale*proportion_exceptions, fill = as.factor(-theta)), position = "dodge") +
  facet_grid(sd_brown ~ sd_tall, scales="free") +
  ylab("difference in mean success probability  /\n proportion of exceptions") +
  xlab("number of objects in context")  + #scale_fill_brewer(palette="Set1", name = "semantic threshold") +
  geom_hline( yintercept = 0) +
  expand_limits(y = c(-0.35*reverse_scale,0.18) ) +
  scale_y_continuous( minor_breaks=c(-0.3*reverse_scale, -0.25*reverse_scale, -0.2*reverse_scale, -0.15*reverse_scale, -0.1*reverse_scale, -0.05*reverse_scale,
                                     0.025, 0.05, 0.075, 0.1, 0.125, 0.15, 0.175),
                      breaks = c(-0.3*reverse_scale,-0.2*reverse_scale,-0.1*reverse_scale,0.05,0,0.1,0.15),
                      labels = c("30\\%", "20\\%","10\\%","0.05","0","0.1","0.15")) +
  scale_fill_manual(values=c("#7a7a7a","#aa8989","#cdcdff","#89aa89","#00aa00","#0000ff","#aa0000","#000000"), breaks = c("0.2", "0.4", "0.6","0.8"), name = "semantic threshold") +
  theme_bw() +
  theme(legend.position="bottom")

show(combined_plot)
```


# Latex 

```{r}
library(tikzDevice)
scale_factor = 0.6
options(tikzLatexPackages=c(getOption("tikzLatexPackages"),"\\usepackage{fixltx2e} \\usepackage{amsmath}"))
tikz("../writing/2019 CogSci/plots/tikz_combined.tex",standAlone=TRUE, width = 9 * scale_factor, height = 7 * scale_factor)
combined_plot
dev.off()

#options(tikzLatexPackages=c(getOption("tikzLatexPackages"),"\\usepackage{fixltx2e} \\usepackage{amsmath}"))
#tikz("../writing/2019 CogSci/plots/tikz_success.tex",standAlone=TRUE, width = 9 * scale_factor, height = 7 * scale_factor)
#results_plot
#dev.off()

#options(tikzLatexPackages=c(getOption("tikzLatexPackages"),"\\usepackage{fixltx2e} \\usepackage{amsmath}"))
#tikz("../writing/2019 CogSci/plots/tikz_exceptions.tex",standAlone=TRUE, width = 9 * scale_factor, height = 7 * scale_factor)
#exceptions_plot 
#dev.off()

```


# Compile

system("pdflatex tikz.tex")
