==> code/08_showcase_graphs.Rmd <==
---
title: "Results Notebook"
output:
  html_document:
    df_print: paged
---

<!--

*Run* by pressing *Ctrl+Shift+Enter*. 
*Insert Chunk* by pressing *Ctrl+Alt+I*.

-->


```{r}
library(scales)
library(tidyverse)
```

number of iterations per bar: 10000


```{r}
selection = read.csv("../results/06_showcase_results.csv") 
selection$success <- selection$tall_brown - selection$brown_tall
```

# Combined graph


```{r}
reverse_scale = 0.25

levels(selection$sd_brown) <- c("$\\sigma_{brown} = 0.1$","$\\sigma_{brown} = 0.2$")
levels(selection$sd_tall) <- c("$\\sigma_{big} = 0.1$","$\\sigma_{big} = 0.2$")

combined_plot = selection %>%
  ggplot(aes(x = as.factor(n_obj))) +
  geom_col(aes(y = success, fill = as.factor(theta)), position = "dodge") +
  geom_col(aes(y = -reverse_scale*proportion_exceptions, fill =
as.factor(-theta)), position = "dodge") +
  facet_grid(sd_brown ~ sd_tall) +
  ylab("difference in mean success probability  /\n proportion of exceptions \n") +
  xlab("\n number of objects in context")  + #scale_fill_brewer(palette="Set1", name = "semantic threshold") +
  geom_hline( yintercept = 0) +
  scale_y_continuous(breaks = c(-0.2*reverse_scale,-0.1*reverse_scale,0.05,0,0.1,0.2), labels = c("20%","10%","0.05","0","0.1","0.2")) +
  scale_fill_manual(values=c("#7a7a7a","#aa8989","#cdcdff","#89aa89","#00aa00","#0000ff","#aa0000","#000000"), breaks = c("0.2", "0.4", "0.6","0.8"), name = "semantic threshold") +
  theme_bw() +
  theme(legend.position="bottom")

show(combined_plot)
```





```{r}
#scale_factor = 0.6
#ggsave(
#  plot = combined_plot, 
#  # this assumes that the CWD is source location!
#  filename = "../writing/2019 CogSci/plots/showcase_results_combined.pdf",
#  height = 7 * scale_factor, 
#  width = 9 * scale_factor
#)
```


```{r}
library(tikzDevice)
## add a package to the defaults
options(tikzLatexPackages=
            c(getOption("tikzLatexPackages"),"\\usepackage{fixltx2e}"))
tikz("tikz.tex",standAlone=TRUE)
library("ggplot2"); theme_set(theme_bw())
reverse_scale = 0.25

levels(selection$sd_brown) <- c("$\\sigma_{brown} = 0.1$","$\\sigma_{brown} = 0.2$")
levels(selection$sd_tall) <- c("$\\sigma_{big} = 0.1$","$\\sigma_{big} = 0.2$")

combined_plot = selection %>%
  ggplot(aes(x = as.factor(n_obj))) +
  geom_col(aes(y = success, fill = as.factor(theta)), position = "dodge") +
  geom_col(aes(y = -reverse_scale*proportion_exceptions, fill =
as.factor(-theta)), position = "dodge") +
  facet_grid(sd_brown ~ sd_tall) +
  ylab("difference in mean success probability  /\n proportion of exceptions \n") +
  xlab("\n number of objects in context")  + #scale_fill_brewer(palette="Set1", name = "semantic threshold") +
  geom_hline( yintercept = 0) +
  scale_y_continuous(breaks = c(-0.2*reverse_scale,-0.1*reverse_scale,0.05,0,0.1,0.2), labels = c("20\\%","10\\%","0.05","0","0.1","0.2")) +
  scale_fill_manual(values=c("#7a7a7a","#aa8989","#cdcdff","#89aa89","#00aa00","#0000ff","#aa0000","#000000"), breaks = c("0.2", "0.4", "0.6","0.8"), name = "semantic threshold") +
  theme_bw() +
  theme(legend.position="bottom")

#p <- ggplot(mpg, aes(x=cty, y=hwy)) + geom_point() +
#  scale_x_continuous(name="text\\textsubscript{subscript}")
#p + annotate("text", x=10, y=40, label="text\\textsubscript{subscript}")
dev.off()

system("pdflatex tikz.tex")
```


# TODO
* Export to Latex
